fix: move upgrade dialog to the RouterSelector.web.js
